An RGB Colour game
I created this web page when following one of the tutorials in the web development courses that I did published by Colt Steele. I wanted to share this because it is one of my favourite web pages that I have created as it looks good, uses techniques which I had not thought of before and taught me a number of programming techniques.